+++ lib/vgs_cg.c
@@ -132,3 +132,26 @@ void vgs_sprite(uint16_t n, BOOL visible, int16_t x, int16_t y, uint8_t size, ui
     OAM[n].rotate = 0;
     OAM[n].scale = 0;
 }
+
+void vgs_pfont_print(uint8_t n, int32_t x, int32_t y, uint8_t pal, uint16_t ptn, const char* text)
+{
+    int32_t dx, dy, width;
+    for (; *text && x < VRAM_WIDTH; text++) {
+        vgs_pfont_get((uint8_t)*text, &dx, &dy, &width);
+        if (0 < width) {
+            vgs_draw_character(n, x + dx, y + dy, FALSE, pal, ptn + *text);
+            x += width;
+        }
+    }
+}
+
+int32_t vgs_pfont_strlen(const char* text)
+{
+    int32_t result = 0;
+    int32_t dx, dy, width;
+    for (; *text; text++) {
+        vgs_pfont_get((uint8_t)*text, &dx, &dy, &width);
+        result += width;
+    }
+    return 0 < result ? result - 1 : result;
+}